feat(grpc): added test profiles to grpc transcoding and fixed dependencies
