+++ integration-tests/grpc-transcoding/src/main/resources/application.properties
@@ -1,20 +1,9 @@
-quarkus.grpc.server.port=9001
-quarkus.grpc.server.instances=2
-quarkus.grpc.server.use-separate-server=false
 quarkus.grpc.transcoding.enabled=true
-
-%vertx.quarkus.grpc.server.use-separate-server=false
-%n2o.quarkus.grpc.server.use-separate-server=true
-%o2n.quarkus.grpc.server.use-separate-server=false
+quarkus.grpc.server.port=9001
 
 quarkus.grpc.clients.hello.host=localhost
 quarkus.grpc.clients.hello.port=9001
 
 %vertx.quarkus.grpc.clients.hello.port=8081
 %vertx.quarkus.grpc.clients.hello.use-quarkus-grpc-client=true
-
-%n2o.quarkus.grpc.clients.hello.port=9001
-%n2o.quarkus.grpc.clients.hello.use-quarkus-grpc-client=true
-
-%o2n.quarkus.grpc.clients.hello.port=8081
-%o2n.quarkus.grpc.clients.hello.use-quarkus-grpc-client=false
+%vertx.quarkus.grpc.server.use-separate-server=false
