+++ integration-tests/grpc-transcoding/src/test/java/io/quarkus/grpc/examples/hello/VertxHelloWorldTranscodingEndpointTest.java
@@ -1,0 +1,10 @@
+package io.quarkus.grpc.examples.hello;
+
+import io.quarkus.grpc.test.utils.VertxGRPCTestProfile;
+import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
+
+@QuarkusTest
+@TestProfile(VertxGRPCTestProfile.class)
+public class VertxHelloWorldTranscodingEndpointTest extends HelloWorldTranscodingEndpointTestBase {
+}
